Fix bugs on counter
